Move server-side relay auth out of shared and drop the unused validator
